requirements.txt
tfx[kfp]==1.12.0
tensorflow==2.11.0
tensorflow_cloud
google-cloud-aiplatform